requirements.txt
absl-py==0.7.1
astor==0.7.1
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.5
emnist==0.0
fabric==2.5.0
future==0.18.2
gast==0.3.3
google-pasta==0.2.0
grpcio==1.23.0
h5py==2.9.0
hyperopt==0.2.3
idna==2.8
invoke==1.4.1
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.3.0
mock==3.0.5
networkx==2.2
nibabel==3.1.0
nltk==3.4.5
numpy==1.19.1
opt-einsum==3.1.0
packaging==20.3
pandas==0.25.1
paramiko==2.7.1
Pillow==7.0.0
protobuf==3.8.0
psycopg2==2.7.6.1
pycparser==2.19
PyNaCl==1.3.0
pyparsing==2.4.2
python-dateutil==2.8.0
python-dotenv==0.11.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
scikit-learn==0.23.0
scipy==1.2.1
simplejson==3.17.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.18
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.10.1
termcolor==1.1.0
threadpoolctl==2.0.0
tqdm==4.42.0
urllib3==1.25.6
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.12.1
